Solve issue with executing plugin in shallow cloned Git repo
The Git lib JGit throws a `JGitInternalException` if `Git#describe()` is
executed in a Git repository, which was cloned shallowly. This behavior
is expected as Git cannot find any information in the local history.

The 'gitDescribeInfo' property is set to 'NA' already if it has no
value. Just catching the exception without any further action should
therefore be sufficient.